# [master] Prioritize DeserializeDelta call for AccountStore