--- a/src/libData/AccountData/AccountStore.cpp
+++ b/src/libData/AccountData/AccountStore.cpp
@@ -34,7 +34,7 @@ using namespace dev;
 using namespace boost::multiprecision;
 using namespace Contract;
 
-AccountStore::AccountStore() {
+AccountStore::AccountStore() : m_externalWriters{0} {
   m_accountStoreTemp = make_unique<AccountStoreTemp>(*this);
 
   if (ENABLE_SC && (!LOOKUP_NODE_MODE || ISOLATED_SERVER)) {
@@ -85,6 +85,8 @@ void AccountStore::InitSoft() {
   unique_lock<shared_timed_mutex> g(m_mutexPrimary);
 
   AccountStoreTrie<unordered_map<Address, Account>>::Init();
+
+  m_externalWriters = 0;
 
   InitRevertibles();
 
@@ -221,7 +223,13 @@ bool AccountStore::DeserializeDelta(const bytes& src, unsigned int offset,
       return false;
     }
   } else {
+    if (LOOKUP_NODE_MODE) {
+      IncrementPrimaryWriteAccessCount();
+    }
     unique_lock<shared_timed_mutex> g(m_mutexPrimary);
+    if (LOOKUP_NODE_MODE) {
+      DecrementPrimaryWriteAccessCount();
+    }
 
     if (!Messenger::GetAccountStoreDelta(src, offset, *this, revertible,
                                          false)) {